refacto(users): move datapass_from_email internal code
