+++ package.json
@@ -72,7 +72,6 @@
     "nprogress": "0.2.0",
     "openid-client": "5.6.4",
     "pg": "8.11.3",
-    "query-string": "8.2.0",
     "static-path": "0.0.4",
     "tailwind-variants": "0.2.0",
     "ts-dedent": "2.2.0",
