+++ packages/~/moderations/lib/package.json
@@ -12,6 +12,7 @@
     }
   },
   "dependencies": {
+    "query-string": "8.2.0",
     "ts-pattern": "5.0.6",
     "zod": "3.22.4"
   },
