+++ packages/~/moderations/lib/src/datapass_from_email.ts
@@ -1,0 +1,16 @@
+//
+
+import queryString from "query-string";
+
+//
+
+export function datapass_from_email(email: string) {
+  const query = queryString.stringify({
+    sorted: JSON.stringify([{ id: "updated_at", desc: false }]),
+    filtered: JSON.stringify([
+      { id: "team_members.email", value: email },
+      { id: "status", value: [] },
+    ]),
+  });
+  return `https://datapass.api.gouv.fr/habilitations?${query}`;
+}
